# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/deletetask": {
            "delete": {
                "description": "Удаляет задачу по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удаление задачи по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача удалена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/gettask": {
            "get": {
                "description": "Возвращает информацию о задаче по её ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получение задачи по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о задаче",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseTask"
                        }
                    }
                }
            }
        },
        "/newtask": {
            "post": {
                "description": "Создает новую задачу и возвращает её ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создание новой задачи",
                "responses": {
                    "202": {
                        "description": "ID созданной задачи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ResponseTask": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "example: 2025-06-25T10:00:00Z",
                    "type": "string"
                },
                "id": {
                    "description": "example: a80f84a8-4841-46a8-bb3f-10d9f775c27a",
                    "type": "string"
                },
                "status": {
                    "description": "example: in_progress",
                    "type": "string"
                },
                "time_duration": {
                    "description": "example: 1h30m",
                    "type": "string"
                }
            }
        }
    }
}